Добавил анбан пользователей

+++ tgbot/bot/modules/contibutions_list/ContributionsList.py
@@ -37,24 +37,38 @@ async def show_list(callback_query: types.CallbackQuery):
     user = await db.Users.find_one({'telegram_id': callback_query.from_user.id})
     region = await db.Regions.find_one({'_id': user['region']})
     status = callback_query.data.split(',')[1]
-    page = 0
-    payments = await get_contributions_list(region['_id'], status, 0)
-    if not payments:
-        return await callback_query.answer('Платежи не найдены')
-
     markup = types.InlineKeyboardMarkup()
-    for payment in payments:
-        payer = await db.Users.find_one({'_id': payment['payer']})
-        fio = f"{payer['second_name']} {payer['first_name'][:1]}.{payer['third_name'][:1]}."
-        markup.add(types.InlineKeyboardButton(text=f"{fio} {payment['payment_date']}", callback_data=f"conlist,payment,{payment['_id']},{page},{status}"))
 
-    if await get_contributions_list(region['_id'], status, 1):
-        button_1 = types.InlineKeyboardButton(text="❌", callback_data=f'conlist,n,0,{status}')
-        button_2 = types.InlineKeyboardButton(text="➡️", callback_data=f'conlist,r,1,{status}')
-        markup.row(button_1, button_2)
+    if status == 'banned':
+        banned_users = await get_banned_list(region['_id'], 0)
+        if not banned_users:
+            return await callback_query.answer('Забаненные пользователи не найдены')
+        for _user in banned_users:
+            fio = f"{_user['second_name']} {_user['first_name'][:1]}.{_user['third_name'][:1]}."
+            markup.add(types.InlineKeyboardButton(text=f"{_user['mention']} - {fio}",
+                                                  callback_data=f"conlist,banned,{_user['_id']},0,{status}"))
+        if await get_banned_list(region['_id'], 1):
+            button_1 = types.InlineKeyboardButton(text="❌", callback_data=f'conlist,n,0,{status}')
+            button_2 = types.InlineKeyboardButton(text="➡️", callback_data=f'conlist,r,1,{status}')
+            markup.row(button_1, button_2)
+        string = 'Выберите пользователя для работы с ним'
+    else:
+        payments = await get_contributions_list(region['_id'], status, 0)
+        if not payments:
+            return await callback_query.answer('Платежи не найдены')
+
+        for payment in payments:
+            payer = await db.Users.find_one({'_id': payment['payer']})
+            fio = f"{payer['second_name']} {payer['first_name'][:1]}.{payer['third_name'][:1]}."
+            markup.add(types.InlineKeyboardButton(text=f"{fio} {payment['payment_date']}", callback_data=f"conlist,payment,{payment['_id']},0,{status}"))
+
+        if await get_contributions_list(region['_id'], status, 1):
+            button_1 = types.InlineKeyboardButton(text="❌", callback_data=f'conlist,n,0,{status}')
+            button_2 = types.InlineKeyboardButton(text="➡️", callback_data=f'conlist,r,1,{status}')
+            markup.row(button_1, button_2)
+        string = 'Выберите платеж для работы с ним'
 
     markup.add(types.InlineKeyboardButton(text='Вернуться', callback_data=f"conlist,back"))
-    string = 'Выберите платеж для работы с ним'
     await callback_query.message.edit_text(string, reply_markup=markup)
 
 
@@ -84,10 +98,49 @@ async def handle_conlist_callback_query(callback_query: types.CallbackQuery):
         callback_query (types.CallbackQuery): Документация на сайте телеграма
     """
 
-    # todo: добавить обработку нажатия на платежи
-    markup = await handle_conlist_callback_query_string_markup_generator(callback_query)
+    split_data = callback_query.data.split(',')
+    if split_data[3] == 'banned':
+        markup = await hande_conlist_banned_callback_query_markup_generator(callback_query)
+    else:
+        markup = await handle_conlist_callback_query_string_markup_generator(callback_query)
     _message = await callback_query.message.edit_reply_markup(reply_markup=markup)
     await callback_query.answer()
+
+
+async def hande_conlist_banned_callback_query_markup_generator(callback_query: types.CallbackQuery):
+    split_data = callback_query.data.split(',')
+    if split_data[1] == 'n':
+        return await callback_query.answer(text='Там больше ничего нет...')
+    db = SingletonClient.get_data_base()
+    user = await db.Users.find_one({'telegram_id': callback_query.from_user.id})
+    page = int(split_data[2])
+    status = split_data[3]
+    region_id = user["region"]
+    logger.info(f"banned list region_id {region_id} page {page}")
+    banned_users = await get_banned_list(region_id, page)
+    markup = types.InlineKeyboardMarkup()
+    for _user in banned_users:
+        fio = f"{_user['second_name']} {_user['first_name'][:1]}.{_user['third_name'][:1]}."
+        markup.add(types.InlineKeyboardButton(text=f"{_user['mention']} - {fio}",
+                                              callback_data=f"conlist,banned,{_user['_id']},0,{status}"))
+    users_l = await get_banned_list(region_id, page - 1)
+    users_r = await get_banned_list(region_id, page + 1)
+    if users_r or users_l:
+        if users_l:
+            left_button = types.InlineKeyboardButton(
+                text='⬅️', callback_data=f'conlist,l,{page - 1},{status}')
+        else:
+            left_button = types.InlineKeyboardButton(
+                text='❌', callback_data=f'conlist,n,{page},{status}')
+        if users_r:
+            right_button = types.InlineKeyboardButton(
+                text="➡️", callback_data=f'conlist,r,{page + 1},{status}')
+        else:
+            right_button = types.InlineKeyboardButton(
+                text='❌', callback_data=f'conlist,n,{page},{status}')
+        markup.row(left_button, right_button)
+    markup.add(types.InlineKeyboardButton(text='Вернуться', callback_data=f"conlist,back"))
+    return markup
 
 
 async def handle_conlist_callback_query_string_markup_generator(callback_query: types.CallbackQuery):
@@ -175,16 +228,65 @@ async def handler_payment_callback(callback_query: types.CallbackQuery):
     await callback_query.message.delete()
 
 
+@dp.callback_query_handler(lambda callback_query: callback_query.data.startswith('conlist,banned'))
+async def handler_banned_callback(callback_query: types.CallbackQuery):
+    logger.info(f'handle conlist banned from {callback_query.from_user.id} data {callback_query.data}')
+    await callback_query.answer()
+    data = callback_query.data.split(',')
+    page = data[3]
+    status = data[4]
+    user_id = ObjectId(data[2])
+    db = SingletonClient.get_data_base()
+
+    user = await db.Users.find_one({
+        '_id': user_id
+    })
+
+    string = f'Пользователь {user.get("second_name")} ' \
+             f'{user.get("first_name")} {user.get("third_name")} ({user.get("mention")})\n'
+
+    button_1 = types.InlineKeyboardButton(text='🥺 Разбанить пользователя',
+                                          callback_data=f'conlist-unban,{user["_id"]},{page},{status}')
+    markup = types.InlineKeyboardMarkup()
+    markup.add(button_1)
+    markup.add(types.InlineKeyboardButton(text='Вернуться',
+                                          callback_data=f"conlist-back,{user['_id']},{page},{status}"))
+
+    await callback_query.message.edit_text(text=string, reply_markup=markup)
+
+
+@dp.callback_query_handler(lambda callback_query: callback_query.data.startswith('conlist-unban'))
+async def handle_conlist_unban_callback(callback_query: types.CallbackQuery):
+    await callback_query.answer()
+    user_id = callback_query.data.split(',')[1]
+    logger.info(f"unban user {user_id}")
+    db = SingletonClient.get_data_base()
+    await db.Users.update_one({'_id': ObjectId(user_id)}, {
+        "$set": {
+            "ban": False
+        }
+    })
+    string = 'Выберите пользователя для работы с ним'
+    markup = await hande_conlist_banned_callback_query_markup_generator(callback_query)
+    _message = await callback_query.message.edit_text(text=string, reply_markup=markup)
+
+
 @dp.callback_query_handler(lambda callback_query: callback_query.data.startswith('conlist-'))
 async def handle_conlist_callback(callback_query: types.CallbackQuery):
     await callback_query.answer()
     if not callback_query.data.startswith('conlist-back'):
         await handle_payment_callback_func(callback_query)
 
-    string = 'Выберите платеж для работы с ним'
-    markup = await handle_conlist_callback_query_string_markup_generator(callback_query)
-    await callback_query.message.answer(string, reply_markup=markup)
-    await callback_query.message.delete()
+    status = callback_query.data.split(',')[3]
+    if status == 'banned':
+        string = 'Выберите пользователя для работы с ним'
+        markup = await hande_conlist_banned_callback_query_markup_generator(callback_query)
+        _message = await callback_query.message.edit_text(text=string, reply_markup=markup)
+    else:
+        string = 'Выберите платеж для работы с ним'
+        markup = await handle_conlist_callback_query_string_markup_generator(callback_query)
+        await callback_query.message.answer(string, reply_markup=markup)
+        await callback_query.message.delete()
 
 
 async def get_contributions_list(region_id, status, page):
@@ -199,5 +301,21 @@ async def get_contributions_list(region_id, status, page):
 
     try:
         return contrib_list[page * 5: page * 5 + 5]
+    except IndexError:
+        return []
+
+
+async def get_banned_list(region_id, page):
+    db = SingletonClient.get_data_base()
+    logger.info(f"get banned list region_id {region_id} page {page}")
+    banned_cursor = db.Users.find({
+        "region": ObjectId(region_id),
+        "ban": True
+    })
+
+    banned_list = await banned_cursor.to_list(length=await db.Users.count_documents({}))
+
+    try:
+        return banned_list[page * 2: page * 2 + 2]
     except IndexError:
         return []
